wrapping up testing for the PluginUtil

+++ plugins/plugin.js
@@ -5,6 +5,7 @@ const fs = require('fs')
 	, ncp = require('ncp').ncp
 	, rimraf = require('rimraf')
 	, {format} = require('util')
+	, log = require('../dist/api/util/Logger').default
 
 //PATHS
 const ROOT_DIR = path.join(__dirname, '..')
@@ -23,8 +24,8 @@ const SETTINGS_PAGE_PATH = path.join(CLIENT_DIR, 'views', 'settings-wrapper.html
 const MANIFEST_PATH = path.join(__dirname, 'manifest.txt')
 const CONFIG_MANIFEST_PATH = path.join(__dirname, 'manifest.config.txt')
 const REGEX_MANIFEST_PATH = path.join(__dirname, 'manifest.regex.txt')
-const SPLASH_PAGE_COPY_PATH = format('%s.backup', SPLASH_PAGE_PATH)
-const SETTINGS_PAGE_COPY_PATH = format('%s.backup', SETTINGS_PAGE_PATH)
+const SPLASH_PAGE_COPY_PATH = path.join(CLIENT_DIR, 'splash.backup')
+const SETTINGS_PAGE_COPY_PATH = path.join(CLIENT_DIR, 'views/settings-wrapper.backup')
 
 //STRINGS
 const NAV_STRING = '<plugins_nav_tabs hidden />'
@@ -49,8 +50,10 @@ function copyRecursive(src, dest){
 }
 
 function init(){
-	fs.copyFileSync(SETTINGS_PAGE_PATH, SETTINGS_PAGE_COPY_PATH)
-	fs.copyFileSync(SPLASH_PAGE_PATH, SPLASH_PAGE_COPY_PATH)
+	if(!fs.existsSync(SETTINGS_PAGE_COPY_PATH))
+		fs.copyFileSync(SETTINGS_PAGE_PATH, SETTINGS_PAGE_COPY_PATH)
+	if(!fs.existsSync(SPLASH_PAGE_COPY_PATH))
+		fs.copyFileSync(SPLASH_PAGE_PATH, SPLASH_PAGE_COPY_PATH)
 	fs.writeFileSync(MANIFEST_PATH, 'SCPP PLUGIN MANIFEST:\n')
 	fs.writeFileSync(CONFIG_MANIFEST_PATH, 'SCPP CONFIG MANIFEST:\n')
 	fs.writeFileSync(REGEX_MANIFEST_PATH, 'SCPP REGEX MANIFEST:\n')
@@ -156,7 +159,7 @@ function registerNavbarLink(targetFilepath, navIconFilepath){
 	let content = fs.readFileSync(SPLASH_PAGE_PATH, 'utf8')
 	content = content.replace(NAV_REGEX, navElement)
 	fs.writeFileSync(SPLASH_PAGE_PATH, content, 'utf8')
-	fs.appendFileSync(navElement + '\n')
+	fs.appendFileSync(REGEX_MANIFEST_PATH, navElement + '\n')
 }
 
 function registerConfigSetting(propname, defaultValue){
@@ -167,51 +170,60 @@ function registerConfigSetting(propname, defaultValue){
 }
 
 function uninstall(){
-	console.log('uninstalling plugins')
+	log.verbose('uninstalling plugins')
 	// delete files registered in the manifest
-	console.log('deleting files from the manifest')
+	log.verbose('deleting files from the manifest')
 	
 	let filesToDelete = fs.readFileSync(MANIFEST_PATH, 'utf8').split('\n')
 	
 	filesToDelete.shift() // remove the title line
 	filesToDelete.forEach(absFilepath => {
 		if(fs.existsSync(absFilepath)){
-			console.log('removing %s', absFilepath)
+			log.verbose('removing %s', absFilepath)
 			fs.unlinkSync(absFilepath)
 		}
 		else console.warn('no file exists to unlink: %s', absFilepath)
 	})
 
 	// remove elements from the config file that were injected (should we do this?)
-	console.log('deleting config from the manifest')
+	log.verbose('deleting config from the manifest')
 	let configToDelete = fs.readFileSync(CONFIG_MANIFEST_PATH, 'utf8').split('\n')
 	configToDelete.shift() //remove the title line
 	let configContent = JSON.parse(fs.readFileSync(CONFIG_FILE_PATH, 'utf8'))
 	configToDelete.forEach(prop => { 
-		console.log('deleting config property %s', prop)
+		log.verbose('deleting config property %s', prop)
 		if(configContent.hasOwnProperty(prop))
 			delete configContent[prop]
 	})
 
-	console.log('loading config file backup')
+	log.verbose('loading config file backup')
 	fs.writeFileSync(CONFIG_FILE_PATH, JSON.stringify(configContent, null, 4), 'utf8')
 	
 	// reinstate backups
 	if(fs.existsSync(SPLASH_PAGE_COPY_PATH)){
-		console.log('loading splash page backup')
+		log.verbose('loading splash page backup')
 		fs.copyFileSync(SPLASH_PAGE_COPY_PATH, SPLASH_PAGE_PATH)
-		console.log('removing splash backup')
+		log.verbose('removing splash backup')
 		fs.unlinkSync(SPLASH_PAGE_COPY_PATH)
 	}
 
 	if(fs.existsSync(SETTINGS_PAGE_COPY_PATH)){
-		console.log('loading settings page backup')
+		log.verbose('loading settings page backup')
 		fs.copyFileSync(SETTINGS_PAGE_COPY_PATH, SETTINGS_PAGE_PATH)
-		console.log('removing settings backup')
+		log.verbose('removing settings backup')
 		fs.unlinkSync(SETTINGS_PAGE_COPY_PATH)
 	}
 
-	// 
+	// cleanup
+	if(fs.existsSync(REGEX_MANIFEST_PATH))
+		fs.unlinkSync(REGEX_MANIFEST_PATH)
+
+	if(fs.existsSync(CONFIG_MANIFEST_PATH))
+		fs.unlinkSync(CONFIG_MANIFEST_PATH)
+
+	if(fs.existsSync(MANIFEST_PATH))
+		fs.unlinkSync(MANIFEST_PATH)
+	
 }
 
 function createEmptyPlugin(name){
